fix(a11y): fix issues
